Completed Homepage
I completed the hover image feature, altered the text in each hover component as well as the introduction and incorporated a map. I also attempted in making a footer which still requires some work. The banner was changed quite a bit due to complications with quality, but for now I'm sticking to the GIF. Then, I organised all my files into folders (which caused some problems with the navigation) and began my second page, where I finished the banner and started on the menu.